backend/requirements.txt
grpcio==1.24.3
grpcio-tools==1.24.3
passlib==1.7.1
protobuf==3.10.0
PyJWT==1.7.1
six==1.12.0
